# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Persona Debate Local API
  version: 1.0.0
  description: |
    Local-first Node.js API for persona/debate/chat/governance workflows.
    All responses use a standard envelope:
    - success: { ok: true, data: ... }
    - error: { ok: false, error: { code, message, details } }
servers:
  - url: http://localhost:3000
security:
  - bearerAuth: []
  - cookieAuth: []
  - apiKeyAuth: []
...
paths:
...
  /api/debates:
    post:
      tags: [Debates]
      summary: Create and queue debate run
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateDebateRequest' }
            examples:
              withManualPersonas:
                value:
                  topic: Should we adopt a four-day work week?
                  context: Compare productivity and burnout outcomes.
                  selectedPersonas:
                    - type: saved
                      id: big-tex
                  knowledgePackIds: []
                  settings:
                    rounds: 3
                    maxWordsPerTurn: 120
                    moderationStyle: neutral
                    sourceGroundingMode: light
                    model: gpt-4.1-mini
                    temperature: 0.7
                    includeModerator: true
      responses:
        '202':
          description: Queued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
        '400': { $ref: '#/components/responses/ErrorResponse' }
    get:
      tags: [Debates]
      summary: List debates
      responses:
        '200':
          description: Debate list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
  /api/debates/{debateId}:
    get:
      tags: [Debates]
      summary: Get debate metadata and transcript text
      parameters:
        - in: path
          name: debateId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Debate detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
  /api/debates/{debateId}/transcript:
    get:
      tags: [Debates]
      summary: Download transcript markdown
      parameters:
        - in: path
          name: debateId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Markdown transcript
          content:
            text/markdown:
              schema:
                type: string
  /api/debates/{debateId}/chat:
    post:
      tags: [Debates]
      summary: Ask follow-up question grounded in transcript excerpts
      parameters:
        - in: path
          name: debateId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [question]
              properties:
                question: { type: string }
                history:
                  type: array
                  items:
                    type: object
                    properties:
                      role: { type: string, enum: [user, assistant] }
                      content: { type: string }
                model: { type: string, default: gpt-4.1-mini }
                temperature: { type: number, default: 0.3 }
      responses:
        '200':
          description: Answer with citations
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
    get:
      tags: [Debates]
      summary: Get persisted follow-up chat history for debate
      parameters:
        - in: path
          name: debateId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: History
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
...
  /api/admin/debates/{debateId}:
    get:
      tags: [Admin]
      summary: Debate analytics detail
      parameters:
        - in: path
          name: debateId
          required: true
          schema: { type: string }
      responses:
        '200': { description: Detail, content: { application/json: { schema: { $ref: '#/components/schemas/OkResponse' } } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    cookieAuth:
      type: apiKey
      in: cookie
      name: pd_session
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
    OkResponse:
      type: object
      properties:
        ok: { type: boolean, example: true }
        data:
          type: object
          additionalProperties: true
    ErrorResponse:
      type: object
      properties:
        ok: { type: boolean, example: false }
        error:
          type: object
          properties:
            code: { type: string, example: VALIDATION_ERROR }
            message: { type: string }
            details:
              nullable: true
              oneOf:
                - type: 'null'
                - type: array
                  items:
                    type: object
                    properties:
                      path: { type: string }
                      message: { type: string }
    SpeakingStyle:
      type: object
      properties:
        tone: { type: string, default: '' }
        verbosity: { type: string, default: '' }
        quirks:
          type: array
          items: { type: string }
      required: [tone, verbosity, quirks]
    Persona:
      type: object
      required: [id, displayName, systemPrompt, speakingStyle, expertiseTags, biasValues, debateBehavior, knowledgePackIds]
      properties:
        id:
          type: string
          pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'
        displayName: { type: string }
        role: { type: string, default: '' }
        description: { type: string, default: '' }
        systemPrompt: { type: string }
        speakingStyle: { $ref: '#/components/schemas/SpeakingStyle' }
        expertiseTags:
          type: array
          items: { type: string }
        biasValues:
          oneOf:
            - type: array
              items: { type: string }
            - type: string
        debateBehavior: { type: string, default: '' }
        knowledgePackIds:
          type: array
          items: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    AdHocPersona:
      allOf:
        - $ref: '#/components/schemas/Persona'
      required: [displayName, systemPrompt, speakingStyle, expertiseTags, biasValues, debateBehavior, knowledgePackIds]
      properties:
        id:
          type: string
          pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'
    SelectedPersona:
      oneOf:
        - type: object
          required: [type, id]
          properties:
            type: { type: string, enum: [saved] }
            id: { type: string }
        - type: object
          required: [type, persona]
          properties:
            type: { type: string, enum: [adhoc] }
            savePersona: { type: boolean, default: false }
            persona: { $ref: '#/components/schemas/AdHocPersona' }
    TopicSource:
      type: object
      required: [title, url]
      properties:
        title: { type: string }
        source: { type: string, default: '' }
        url: { type: string, format: uri }
        publishedAt:
          oneOf:
            - type: string
            - type: 'null'
        snippet: { type: string, default: '' }
    DebateSettings:
      type: object
      properties:
        rounds: { type: integer, minimum: 1, maximum: 8, default: 3 }
        maxWordsPerTurn: { type: integer, minimum: 40, maximum: 400, default: 120 }
        moderationStyle: { type: string, default: neutral }
        sourceGroundingMode: { type: string, enum: [off, light, strict], default: light }
        model: { type: string, default: gpt-4.1-mini }
        temperature: { type: number, minimum: 0, maximum: 2, default: 0.7 }
        includeModerator: { type: boolean, default: true }
    CreateDebateRequest:
      type: object
      required: [topic]
      properties:
        topic: { type: string }
        context: { type: string, default: '' }
        selectedPersonas:
          type: array
          items: { $ref: '#/components/schemas/SelectedPersona' }
          default: []
        settings: { $ref: '#/components/schemas/DebateSettings' }
        knowledgePackIds:
          type: array
          items: { type: string }
          default: []
        topicDiscovery:
          type: object
          properties:
            query: { type: string, default: '' }
            selectedTitle: { type: string, default: '' }
            selectedSummary: { type: string, default: '' }
            sources:
              type: array
              items: { $ref: '#/components/schemas/TopicSource' }
...
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'